Exercises/Exercise-9/requirements.txt
polars-lts-cpu==1.6.0
